Allow for attributes that start with any verb to fix grammar of Pet attribute
This change moves the attribute verb e.g. is, owns from the fact pattern to the attribute description
to allow for different verbs for different attribute. The PersonName Attribute
remains handles specially because it doesn't contain a verb.

In a follow up CL, the special handling of PersonName will be replaced with full support for
noun attributes like "The Frenchman" from the original zebra puzzle.